program the project

+++ ch07/sum2.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 
 int main(void) {
-  long n, sum = 0;
+  double n, sum = 0;
 
   printf("This program sums a series of integers.\n");
   printf("Enter integers (0 to terminate): ");
 
   do {
-    scanf("%ld", &n);
+    scanf("%lf", &n);
     sum += n;
   } while (n != 0);
 
-  printf("The sum is: %ld\n", sum);
+  printf("The sum is: %lf\n", sum);
 
   return 0;
 }
